refactor: renamed chipmenu
